src/tests/requirements.txt
asgiref==3.9.1
attrs==25.3.0
bcrypt==4.1.3
black==24.8.0
cfgv==3.4.0
click==8.2.1
colorama==0.4.6
distlib==0.4.0
Django==5.0.7
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2025.9.1
filelock==3.19.1
identify==2.6.14
inflection==0.5.1
iniconfig==2.1.0
isort==5.13.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
pre-commit==3.8.0
psycopg==3.2.1
psycopg-binary==3.2.1
PyJWT==2.9.0
pytest==8.3.2
pytest-django==4.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.1
ruff==0.6.2
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
virtualenv==20.34.0
